src/005-Communication & Visualization (Days 51-60)/day_058_collaborative_analysis_notebook/main.py
import os
import sys
from dotenv import load_dotenv

def main():
    """
    Main entry point for Day 058 Collaborative Analysis Notebook.
    """
    # Load environment variables from .env file
    # This looks for a .env file in the current directory or parents
    load_dotenv()

    # Securely access the API key
    # Modify this to match the specific provider you are using (e.g., ANTHROPIC_API_KEY, GEMINI_API_KEY)
    api_key = os.getenv("OPENAI_API_KEY")

    if not api_key:
        print("WARNING: OPENAI_API_KEY not found in environment variables.")
        print("Please ensure you have a .env file in the project root or this directory.")
        # sys.exit(1) # Optional: Exit if critical
    else:
        print("Successfully loaded API key from environment.")

    print("-" * 50)
    print(f"Starting Project: {os.path.basename(os.getcwd())}")
    print("-" * 50)

    # TODO: Implement Day 058 Collaborative Analysis Notebook logic here

    print("Project setup complete. Ready for logic implementation.")

if __name__ == "__main__":
    main()
