requirements.txt
# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
polars>=0.19.0

# Deep Learning (Select based on preference, including both for completeness)
torch>=2.0.0
tensorflow>=2.14.0

# AI & LLMs
openai>=1.0.0
anthropic>=0.3.0
langchain>=0.1.0
langchain-openai
langchain-anthropic

# Environment Management
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
streamlit>=1.30.0

# MLOps & Tracking
mlflow>=2.9.0
wandb>=0.16.0

# Ethics & Interpretability
shap>=0.44.0
fairlearn>=0.10.0
interpret>=0.5.0

# Utilities
tqdm>=4.66.0
requests>=2.31.0
jupyter>=1.0.0
